update pytest requirements
file structure
