+++ app/__init__.py
@@ -1,0 +1,27 @@
+
+#
+# this is the app/__init__.py file
+#
+# the presence of this file, even if empty,
+# ... facilitates local imports from one file to another,
+# ... including importing app code for testing purposes
+#
+#
+
+
+def function(x):
+    return x+x
+
+
+
+
+#def function(my_price):
+#    """
+#    Formats a number as currency (USD) with dollar sign, two decimals, and thousands separator
+#    
+#    Params: my_price (numeric, int or float) to be formatted
+#    
+#    Examples: to_usd(21.645)
+#    """
+#    return f"${my_price:,.2f}"
+#
